+++ mpi/init_eqlbrmdistrfuncDF0.c
@@ -41,33 +41,32 @@ void iniEquilibrium(Fluidnode* node) {
 void init_eqlbrmdistrfuncDF0(GV gv){/*stored in dfeq*/
   Fluidgrid     *fluidgrid;
 
-  long starting_x, starting_y, starting_z, stopping_x, stopping_y, stopping_z;//To identify buffer zone
-
+  int starting_x, starting_y, starting_z, stopping_x, stopping_y, stopping_z;//To identify buffer zone
 
   fluidgrid = gv->fluid_grid;
 
   int cube_size = gv->cube_size;
-  long num_cubes_x = gv->fluid_grid->num_cubes_x;
-  long num_cubes_y = gv->fluid_grid->num_cubes_y;
-  long num_cubes_z = gv->fluid_grid->num_cubes_z;
+  int num_cubes_x = gv->fluid_grid->num_cubes_x;
+  int num_cubes_y = gv->fluid_grid->num_cubes_y;
+  int num_cubes_z = gv->fluid_grid->num_cubes_z;
   int  tmp_task;
 
   /*PTHREAD_Change*/
 
   //For each cube BI, BJ ,BK, li, lj, lk 0 to cube size-1
-  for (long BI = 0; BI < num_cubes_x; ++BI){
-    for (long BJ = 0; BJ < num_cubes_y; ++BJ){
-      for (long BK = 0; BK < num_cubes_z; ++BK){
+  for (int BI = 0; BI < num_cubes_x; ++BI){
+    for (int BJ = 0; BJ < num_cubes_y; ++BJ){
+      for (int BK = 0; BK < num_cubes_z; ++BK){
         tmp_task = cube2task(BI, BJ, BK, gv);
         if (gv->taskid == tmp_task){ //MPI changes
-          long cube_idx = BI * num_cubes_y * num_cubes_z + BJ * num_cubes_z + BK;
+          int cube_idx = BI * num_cubes_y * num_cubes_z + BJ * num_cubes_z + BK;
           Fluidnode *nodes = fluidgrid->sub_fluid_grid[cube_idx].nodes;
           starting_x = starting_y = starting_z = 0;
           stopping_x = stopping_y = stopping_z = cube_size - 1;
-          for (long li = starting_x; li <= stopping_x; ++li){
-            for (long lj = starting_y; lj <= stopping_y; ++lj){
-              for (long lk = starting_z; lk <= stopping_z; ++lk){
-                long node_idx = li * cube_size * cube_size + lj * cube_size + lk; //local node index inside a cube.
+          for (int li = starting_x; li <= stopping_x; ++li){
+            for (int lj = starting_y; lj <= stopping_y; ++lj){
+              for (int lk = starting_z; lk <= stopping_z; ++lk){
+                int node_idx = li * cube_size * cube_size + lj * cube_size + lk; //local node index inside a cube.
                 iniEquilibrium(nodes+node_idx);
 
                 // for (int ksi = 0; ksi<19; ++ksi){
